customer service: service implementation work in progress

+++ customer/src/main/java/fi/koku/services/entity/customer/impl/CustomerQueryCriteria.java
@@ -1,6 +1,5 @@
 package fi.koku.services.entity.customer.impl;
 
-import java.math.BigInteger;
 
 /**
  * Customer query criteria.
@@ -8,17 +7,17 @@ import java.math.BigInteger;
  * @author aspluma
  */
 public class CustomerQueryCriteria {
-  private BigInteger id;
+  private Long id;
   private String pic;
   private String selection;
 
-  public CustomerQueryCriteria(BigInteger id, String pic, String selection) {
+  public CustomerQueryCriteria(Long id, String pic, String selection) {
     this.id = id;
     this.pic = pic;
     this.selection = selection;
   }
 
-  public BigInteger getId() {
+  public Long getId() {
     return id;
   }
